# clean_website/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Study Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Not found
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Study Materials ───────────────────────────────────────────────────────
...
  /materials/{id}/exams:
    get:
      operationId: listExams
      tags: [exams]
      summary: List exams for a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
    post:
      operationId: generateExam
      tags: [exams]
      summary: Generate an exam from material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamRequest"
      responses:
        "201":
          description: Generated exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /exams/{id}:
    get:
      operationId: getExam
      tags: [exams]
      summary: Get an exam with questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
    delete:
      operationId: deleteExam
      tags: [exams]
      summary: Delete an exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /exams/{id}/submit:
    post:
      operationId: submitExam
      tags: [exams]
      summary: Submit exam answers for grading
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamSubmission"
      responses:
        "200":
          description: Graded result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResult"
  /exam-results/{id}:
    get:
      operationId: getExamResult
      tags: [exams]
      summary: Get a graded exam result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResult"

  # ── AI Chat ───────────────────────────────────────────────────────────────
  /materials/{id}/chat:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get chat history for a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the AI tutor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Course:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        university:
          type: ["string", "null"]
        semester:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        materialCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CourseInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        university:
          type: string
        semester:
          type: string
        color:
          type: string
    CourseUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        university:
          type: string
        semester:
          type: string
        color:
          type: string

    # ── Material ─────────────────────────────────────────────────────────────
...
    Question:
      type: object
      required: [id, setId, questionType, question, answer, createdAt]
      properties:
        id:
          type: integer
        setId:
          type: integer
        questionType:
          type: string
          enum: [open, multiple_choice, true_false, fill_blank, short_answer]
        question:
          type: string
        answer:
          type: string
        explanation:
          type: ["string", "null"]
        options:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        createdAt:
          type: string
          format: date-time
...
    Exam:
      type: object
      required: [id, materialId, title, language, examType, questionCount, createdAt]
      properties:
        id:
          type: integer
        materialId:
          type: integer
        title:
          type: string
        language:
          type: string
          enum: [he, en]
        examType:
          type: string
          enum: [practice, topic_quiz, midterm, final]
        questionCount:
          type: integer
        timeLimitMinutes:
          type: ["integer", "null"]
        difficulty:
          type: string
          enum: [easy, medium, hard, mixed]
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        createdAt:
          type: string
          format: date-time
    ExamRequest:
      type: object
      required: [language, examType]
      properties:
        language:
          type: string
          enum: [he, en]
        examType:
          type: string
          enum: [practice, topic_quiz, midterm, final]
        questionCount:
          type: integer
        timeLimitMinutes:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard, mixed]
        topics:
          type: array
          items:
            type: string
    ExamSubmission:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/ExamAnswer"
        timeSpentSeconds:
          type: integer
    ExamAnswer:
      type: object
      required: [questionId, answer]
      properties:
        questionId:
          type: integer
        answer:
          type: string
    ExamResult:
      type: object
      required: [id, examId, score, totalQuestions, correctCount, createdAt]
      properties:
        id:
          type: integer
        examId:
          type: integer
        score:
          type: number
        totalQuestions:
          type: integer
        correctCount:
          type: integer
        timeSpentSeconds:
          type: ["integer", "null"]
        feedback:
          type: array
          items:
            $ref: "#/components/schemas/QuestionFeedback"
        createdAt:
          type: string
          format: date-time
    QuestionFeedback:
      type: object
      required: [questionId, correct, userAnswer, correctAnswer]
      properties:
        questionId:
          type: integer
        correct:
          type: boolean
        userAnswer:
          type: string
        correctAnswer:
          type: string
        explanation:
          type: ["string", "null"]

    # ── Chat ─────────────────────────────────────────────────────────────────
    ChatMessage:
      type: object
      required: [id, materialId, role, content, createdAt]
      properties:
        id:
          type: integer
        materialId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        language:
          type: ["string", "null"]
          enum: [he, en, mixed]
        createdAt:
          type: string
          format: date-time
    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        language:
          type: string
          enum: [he, en]

    # ── Dashboard ─────────────────────────────────────────────────────────────
...