# clean_website/lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Study Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessageLanguage } from './chatMessageLanguage';
import type { ChatMessageRole } from './chatMessageRole';

export interface ChatMessage {
  id: number;
  materialId: number;
  role: ChatMessageRole;
  content: string;
  /** @nullable */
  language?: ChatMessageLanguage;
  createdAt: Date;
}