# clean_website/lib/api-zod/src/generated/types/examAnswer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Study Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ExamAnswer {
  questionId: number;
  answer: string;
}